refactor(backend): remove graphql module
